+++ gateway/constants.js
@@ -1,1 +1,0 @@
-module.exports.DEFAULT_CONF_FILE = './auth-gateway.yaml';