+++ gateway/server/auditor/null.js
@@ -1,9 +1,0 @@
-const { Auditor } = require('./base');
-
-
-class NullAuditor extends Auditor {
-  audit() {
-    return Promise.resolve(null);
-  }
-};
-module.exports.NullAuditor = NullAuditor;